==> notebooks/NB2_target_location_identification_with_BERT.ipynb <==
# third party imports
import pandas as pd
import numpy as np
from tqdm import tqdm
# local imports
from BERT_geoparser.data import Data, Phrase
from BERT_geoparser.tokenizer import Tokenizer
from BERT_geoparser.model import BertModel
from BERT_geoparser.analysis import Results
from BERT_geoparser.retagger import Retagger


# Notebook 2: building a target location identification model
This notebook goes through the steps neccessary for using a trained NER-tagging model (as built in notebook 1) to build a target/incidental location BERT model. This will involve 3 steps:

1. Open the yelp review dataset and link each review to the geographic location of the business.
2. Use a trained BERT model to NER tag the text in each review.
3. Retag the data by assessing the proximity of any tokens (or strings of tokens) tagged as locations to the true location identified in step 1.
4. Use this re-tagged data to retrain the BERT-model on the task of identifying target and incidental locations mentioned in text.

By doing this, the goal is to produce a model which can parse locations in text in a sensible way - identifying locations which the text is directly referencing ('target' locations) and other locations which are not related to the business being reviewed ('incidental' locations). For example,  the sentence:

 <p style="text-align: center;"> "Donnies Pizza Heaven in <span style="color:green">Inidanapolis</span> do the best deep pan pizza outside of <span style="color:red">Chicago</span>." </p>

 would receive the tags:

  <p style="text-align: center;"> "[O] [O] [O] [O] <span style="color:green">[B-tar]</span> [O] [O] ... <span style="color:red">[B-inc]</span>." </p>

  ## Step 1: adding locations to review data
  The first step is to add the locations to the yelp review data. We can do this by linking the `yelp_academic_dataset_review.json` data with the `yelp_academic_dataset_business.json` data using the `business_id` column.

# Get the review dataset and the accompanying business info dataset
review_df = pd.read_json('../data/yelp_academic_dataset_review.json', lines=True, nrows=25000)#[10000:]
business_df = pd.read_json('../data/yelp_academic_dataset_business.json', lines=True)
#review_df = pd.read_csv('data/model_improvement/randomised_review_locations.csv', nrows=1000)#[10000:]

# add locations to reviews by linking to business
def get_coords(business_ids):
    coords = []
    for id in tqdm(business_ids):
        business = business_df[business_df.business_id==id]
        lat = business.latitude.iloc[0]
        lon = business.longitude.iloc[0]
        coords.append((lat, lon))
    return coords

review_df['coordinates'] = get_coords(review_df.business_id.values)

review_df.to_csv('data/step_2/25k_yelp_reviews_with_location.csv')

## Step 2: parsing the reviews using the BERT model
We will load a pre-trained BERT model using the `Data`, `Tokenizer` and `BERTModel` classes and use this to parse the text in the review dataset. This will output a dataframe set up in a way that makes it easy to use to re-train a new model on in the future. 

# Load the dataset using the BERT_geoparser Data.py module
data_csv = r'../data/ner_dataset.csv'
tokenizer = Tokenizer(size='base', cased=False)
data = Data(data_path=data_csv, 
            tokenizer=tokenizer,
            max_len=125)

model = BertModel(saved_model='20230808_bert_model_large.hdf5', data=data)
model.model.summary()

results = model.results_dataframe(texts=review_df.text.values, include_best=True)

## Step 3: re-tagging the reviews with target/incidental locations
We can now use the `Retagger` class to tag every location identified in step 2 with a new 'target' or 'incidental' tag. We will consider locations as being anything with a tag containing either `geo`, `org` or `gpe`. Everything elase will be given an `O` tag. 

We will complete proximity checks using a bounding box around the matched locations. This is acheived by setting `threshold='bbox'`. This will mean that any tokens or phrase for which the true location is within the bounding box of any locations matched to the word or phrase will be tagged with `tar` and other will be tagged `inc`. Alternatively, we could set a minimum distance (in KM) for these proximity checks by setting `threshold=<float>`.

retagger = Retagger(results)
retagger.retag(['geo', 'gpe', 'org'], threshold='bbox', review_df=review_df)
retagged_data = retagger.df

retagged_data.to_csv('data/step_4/test_data_random_locations.csv', index=False)

# get some info about distribution of targets
O = retagged_data[retagged_data.Tag == 'O']
B_inc = retagged_data[retagged_data.Tag == 'B-inc']
B_tar = retagged_data[retagged_data.Tag == 'B-tar']
I_inc = retagged_data[retagged_data.Tag == 'I-inc']
I_tar = retagged_data[retagged_data.Tag == 'I-tar']

print(f'O tags : {len(O)}')
print(f'B-tar tags : {len(B_tar)}')
print(f'I-tar tags : {len(I_tar)}')
print(f'B-inc tags : {len(B_inc)}')
print(f'I-inc tags : {len(I_inc)}')

## Step 4: Retrain a new BERT-model on the new tags


data_path = r'data/step_4/test_yelp_tar_inc_tagged_bbx_org_geo_gpe.csv'
tokenizer = Tokenizer(size='base', cased=False)
data = Data(data_path=data_path, tokenizer=tokenizer, max_len=125)
tar_model = BertModel(saved_model=False, data=data)
tar_model.model.summary()

tar_model.train(save_as='20230926_tar_tagged_bert_model_large.hdf5', n_epochs=4, batch_size=16, validation_split=0.1)

# Get the review dataset and the accompanying business info dataset
review_df = pd.read_json('../data/yelp_academic_dataset_review.json', lines=True, nrows=25000)[20000:]
business_df = pd.read_json('../data/yelp_academic_dataset_business.json', lines=True)

# add locations to reviews by linking to business
def get_coords(business_ids):
    coords = []
    for id in tqdm(business_ids):
        business = business_df[business_df.business_id==id]
        lat = business.latitude.iloc[0]
        lon = business.longitude.iloc[0]
        coords.append((lat, lon))
    return coords

review_df['coordinates'] = get_coords(review_df.business_id.values)

# Load the dataset using the BERT_geoparser Data.py module
data_csv = r'../data/ner_dataset.csv'
tokenizer = Tokenizer(size='base', cased=False)
data = Data(data_path=data_csv, 
            tokenizer=tokenizer,
            max_len=125)

model = BertModel(saved_model='20230808_bert_model_large.hdf5', data=data)
results = model.results_dataframe(texts=review_df.text.values, include_best=True)

results = results.reset_index()

retagger = Retagger(results)
review_df = review_df.reset_index()
retagger.retag(['geo', 'gpe', 'org'], threshold='bbox', review_df=review_df)
test_data = retagger.df

test_data.to_csv('data/step_4/test_yelp_dataset.csv', index=False)

data_csv = 'data/step_4/test_yelp_dataset.csv'
tokenizer = Tokenizer(size='base', cased=False)
data = Data(data_path=data_csv, 
            tokenizer=tokenizer,
            max_len=125)
tar_model = BertModel(saved_model='20230926_tar_tagged_bert_model_large.hdf5', data=data)
X_tokens, y_pred, y_true = tar_model.test('data/step_4/test_yelp_dataset.csv', return_tokens=True)

res = Results(y_true, y_pred)
for cat in ['O', 'tar', 'inc']:
    print(f'"{cat}" accuracy : {np.round(res.categorical_accuracy(cat),3)}')
    print(f'"{cat}" precision : {np.round(res.categorical_precision(cat),3)}')
    print(f'"{cat}" recall : {np.round(res.categorical_recall(cat),3)}')
    print('=======================')
print(f'macro average recall : {np.round(res.macro_average_recall(), 3)}')
print(f'macro average precision : {np.round(res.macro_average_precision(),3)}')
print(f'micro average recall : {np.round(res.micro_average_recall(),3)}')
print(f'micro average precision : {np.round(res.micro_average_precision(),3)}')

rows = {'Sentence #': [], 'Word':[], 'Tag':[], 'Predicted Tag':[]}
for i, (X, y, y_hat) in enumerate(zip(X_tokens, y_true, y_pred)):
    rows['Sentence #'].extend([i]*len(X))
    rows['Word'].extend(X)
    rows['Tag'].extend(y)
    rows['Predicted Tag'].extend(y_hat)
results_df = pd.DataFrame(rows)

results_df.to_csv('data/predictions/20230927_tar_inc_predictions.csv', index=False)

## Model improvement
While these results are promising, there is a serious flaw in the datset the model has been trained on. Specifically, the reviews are from businesses from a small number of locations. As such, there is a danger that the model is simply consistently assigning the `tar` tag to tokens matching those places. To investigate we will build a new dataset with all the tokens tagged as `tar` replaced with random draws from a list of place names. 

# open the test data
test_data = pd.read_csv('data/step_4/test_yelp_dataset.csv')
# extract only lines tagged as target
target_only_data = test_data[test_data.Tag.str.contains('tar')]
# use the Retagger class to add a 'sequential group' column to this data.
retagger = Retagger(target_only_data)
retagger.add_sequential_groups()
retagger.df.head()

import random
from BERT_geoparser.utils import flatten

world_cities_df = pd.read_csv('data/model_improvement/worldcities.csv')
us_cities_df = world_cities_df[world_cities_df.iso3=='USA']
us_cities = []
for i, city in us_cities_df.iterrows():
    name = city.city_ascii
    lat = city.lat
    lng = city.lng
    us_cities.append((name, (lat,lng)))

replacement_dict = {}

for i, group in retagger.df.groupby('sequential_group'):
    phrase = Phrase('', tag=None)
    for token, tag in zip(group['Word'].values, group['Tag'].values):
        phrase.add_token(token=token, tag=tag)
    if phrase.text not in replacement_dict.keys():
        new_cities = random.choices(us_cities, k=2)
        # make sure we're not replacing with the same city
        new_city = new_cities[0]
        if new_city == phrase.text:
            new_city == new_cities[1]
        replacement_dict.update({phrase.text:new_city})  

# Get the review dataset and the accompanying business info dataset
review_df = pd.read_csv('data/step_2/yelp_reviews_with_location.csv')

print(review_df.iloc[0].text)
print(review_df.iloc[0].coordinates)

# loop over each review and replace any identified target locations using the dictionart
for i, review in review_df.iterrows():
    text = review.text.lower()
    words = text.split(' ')
    new_coordinates = review.coordinates
    for w in words:
        if w in replacement_dict.keys():
            new_text = text.replace(w, replacement_dict[w][0])
            new_coordinates = replacement_dict[w][1]
            review_df.loc[i, 'text'] = new_text
            review_df.loc[i, 'coordinates'] = str(new_coordinates)
       

# Check the first entry: NJ should have been replaced with a new place and the location updated
print(review_df.iloc[0].text)
print(review_df.iloc[0].coordinates)

review_df.to_csv('data/model_improvement/randomised_review_locations.csv', index=False)

data_csv = 'data/step_4/test_data_random_locations.csv'
tokenizer = Tokenizer(size='base', cased=False)
data = Data(data_path=data_csv, 
            tokenizer=tokenizer,
            max_len=125)
tar_model = BertModel(saved_model='20230926_tar_tagged_bert_model_large.hdf5', data=data)
X_tokens, y_pred, y_true = tar_model.test('data/step_4/test_data_random_locations.csv', return_tokens=True)

res = Results(y_true, y_pred)
for cat in ['O', 'tar', 'inc']:
    print(f'"{cat}" accuracy : {np.round(res.categorical_accuracy(cat),3)}')
    print(f'"{cat}" precision : {np.round(res.categorical_precision(cat),3)}')
    print(f'"{cat}" recall : {np.round(res.categorical_recall(cat),3)}')
    print('=======================')
print(f'macro average recall : {np.round(res.macro_average_recall(), 3)}')
print(f'macro average precision : {np.round(res.macro_average_precision(),3)}')
print(f'micro average recall : {np.round(res.micro_average_recall(),3)}')
print(f'micro average precision : {np.round(res.micro_average_precision(),3)}')

tar_model.results_dataframe(texts=["Donnies Pizza Heaven in Derby do the best deep pan pizza outside of Lockport"], include_best=True)

replacement_dict
